requirements.txt
catboost==0.24.3
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.2
future==0.18.2
graphviz==0.15
h2o==3.32.0.2
hyperopt==0.2.5
idna==2.10
joblib==1.0.0
kaleido==0.1.0
kiwisolver==1.3.1
lightgbm==3.1.1
llvmlite==0.35.0
matplotlib==3.3.3
networkx==2.5
numba==0.52.0
numpy==1.19.3
pandas==1.1.5
Pillow==8.0.1
plotly==4.14.1
PDPbox==0.2.0
pyodbc==4.0.30
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
requests-cache==0.5.2
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
shap==0.37.0
six==1.15.0
slicer==0.0.3
threadpoolctl==2.1.0
tqdm==4.54.1
urllib3==1.26.2
xgboost==1.2.1
